move comparatorMin
